Separating TabWidget and TabHeaderWidget into two files. #54
